django_server/requirements.txt
django
psycopg2
PyYAML
django-twilio
django-extensions
celery
mosquitto
django-json-field
django-model-utils
